chara_fur: CalcMogScore signed dist-sum conversion
CalcMogScore 92.11->92.29, unit 90.875->90.895. Wrap the
dx*dx+dy*dy distance sum in static_cast<int> before the double
conversion so MWCC emits the signed int->double path (xoris +
signed bias magic) matching target, while leaving dx unsigned for
the atan2 site. Sum-of-squares is always small positive so the
signed/unsigned double value is identical (behavior-preserving).

+++ src/chara_fur.cpp
@@ -584,7 +584,7 @@ void CChara::CalcMogScore()
 		for (int x = 0; x < 0x40; x++) {
 			const unsigned int dx = x - 0x20;
 			const int dy = y - 0x20;
-			const int dist = static_cast<int>(sqrt(static_cast<double>(dx * dx + dy * dy)));
+			const int dist = static_cast<int>(sqrt(static_cast<double>(static_cast<int>(dx * dx + dy * dy))));
 
 			if (dist >= 0x40) {
 				continue;
